Restrict child position based on parent position

This constraint requires children to be positioned at a certain side of their parent and with a certain distance. It can be used to enforce a layering of particles for e.g. DAG and tree layouts.

Training parameters

The following parameters defines the training of the constraint and can be passed along a call to impose()

distance : The minimum orthogonal distance to the parent. Default to 0, meaning that children are only required to be positioned to the specific side of their parent. (tidy eval)

angle : The direction the children should be enforced to be relative to their parent. Defaults to -pi/2 which is equivalent to down. (tidy eval)

See also