Apply a map to particles

In mathematics, maps are a functions that translates its input into new values. In the context of particles a map is a translation function that translates the current particle positions to a new one

Details

Normally a map has no notion of velocity — it simply translates positions. In particles it is possible to decide whether positions should be modified directly or whether the translation magnitude should be added to the velocity verlet using the fixed parameter.

Training parameters

The following parameters defines the training of the force and can be passed along a call to wield()

map : A function that accepts the particle position matrix and returns the new positions in the same format.

fixed : Logical. Should position be modified directly (TRUE) or should the translation be added to the velocity verlet (FALSE)

See also