Apply the mean velocity of all the neighbors to a particle

This force takes the mean of all the neighbors (in the graph sense) of a particle (and optionally itself) and applies it to itself.

Training parameters

The following parameters defines the training of the force and can be passed along a call to wield()

include_self : Should the velocity of itself be included in the mean calculation

mode : How should neighbors be found? 'all' uses all edges. 'out' only uses outbound edges, and 'in' only uses inbound edges. Ignored for undirected particle graphs

See also