Fixes particles to be inside a polygon

This constraint prevents particles from moving outside of one or more polygons. If a particle ventures outside it will be moved back to its closest point inside the specified polygon(s) and have its velocity set to zero.

Training parameters

The following parameters defines the training of the constraint and can be passed along a call to impose()

polygon : A two column matrix giving the polygon, or a list of matrices to use multiple polygons. Overlapping polygons will be subtracted from each other so it is possible to define polygons with holes.

See also