Attract particles towards a vertical position

This force simply pulls particles towards a fixed position on the y-axis.

Training parameters

The following parameters defines the training of the force and can be passed along a call to wield()

strength : The strength with which the attraction occurs (tidy eval)

y : The position on the y-axis to pull towards. (tidy eval)

See also